+++ src/deltachat_cursed/util.py
@@ -26,6 +26,8 @@ COMMANDS = {
         "/nick",
         "/pin",
         "/unpin",
+        "/mute",
+        "/unmute",
         "//",
     ]
 }